requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
Pillow==10.1.0
openai-whisper==20231117
yt_dlp==2025.12.8
requests==2.31.0
ffmpeg-python==0.2.0
itsdangerous==2.2.0